Fix strange hints issue after revert ec79ba4
The revert ec79ba4 was applied so that `menu_text` didn't get used as
`sort_property` anymore.

The reason for that remotion was that it caused the hint completion to
vanish after typing past the hint trigger (as intended), regardless the
menu itens matching characters being typed . Previously, if what got
typed matched hint menu itens, they didn't vanish, and no useful
"non hint" completion was shown.

But leaving an empty string as `sort_property` at
`FilterAndSortCandidates` seems to have caused strange side effects.

YCM's `ConvertCompletionDataToVimData` started to receive strings as
`completion_data` instead of dictionaries, which caused exceptions to be
thrown when indexing it. It happened on typing after the hint menu got
displayed.

This all got related with commit 9feb587 that completely refactored
`FilterAndSortCandidatesWrap` regarding encoding issues. The new added
function `_ConvertCandidates` was checking `sort_property`, and if it
was empty it called `converter` for the whole `candidate` instead of
just leaving it as it arrived?

Why convert the whole candidate and return it for no sort_property?

No idea. So it's now just returning the original in case `sort_property`
is empty.  It fixes the problem with `ConvertCompletionDataToVimData`
receiving strings instead of objects at the other end.

TODO: It should be investigated whether empty `sort_property` is really
the best/correct choice to make the hint menu to vanish after typing
past the hint trigger.

    void foo() {}
    void bar(float) {}
    void baz(float x) {}

    int main() {
        foo(vo|); // hint menu does not vanish at |
        bar(fl|); // hint menu does not vanish at |
        baz(fl|); // hint menu vanishes at | for named parameters
    }