Python 2 unicode test fixes
Without the custom string caster, `std::string` values become
`unicode`s, which breaks a couple tests.

First, xml.etree.ElementTree.XML under Python 2 doesn't allow `unicode`
to be passed in: if it gets one it attempts to convert it to ascii,
which fails.  But give it a `str` containing encoded utf-8 data and it's
happy.

The second set of test fixes are much more minor: the regex string
matches weren't allowing for a leading `u` in the result representation.