Implement `applyEdit` to allow incomplete `FixIts`
When making an `executeCommand` request, a LSP server might in exchange
send a series of "things", which we expect to be `applyEdit` requests.

We then store those requests in a list until the LSP server responds to
the original `executeCommand` and finally let the specific LSP completer
implementation figure out what to do with those unconsumed `applyEdit`s.

This allows ycmd to forward an incomplete `codeAction` from clangd to a
client in order to save time computing expensive `codeActions`.

The exchange should look like this:

- A client makes a `FixIt` request.
- ycmd makes a `codeAction` request to clangd.
- clangd responds with a mix of resolved and unresolved code actions.
- ycmd transforms resolved code actions in `FixIts` and forwards
unresolved ones.
- A client lets a user choose a fixit from the list.
- If the chosen fixit is unresolved, the client sends a /resolve_fixit
request.
- ycmd executes LSP server specific logic to resolve the fixit and
responds with another `FixItResponse`.
- In case of clangd, the server specific logic is:
  - Make the appropriate `executeCommand` request.
  - Store the received `applyEdit`.
  - Turn the `applyEdit` into a ycmd `FixIt`.
  - Send a `FixItResponse` to the client.

Open questions:

- Should `ResolveFixit` in `ClangdCompleter` be moved to
`SimpleLSPCompleter` or even `LanguageServerCompleter`?
- How should "stray" `applyEdit`s be drained?
- What if an `executeCommand` triggers server-to-client requests other
than `applyEdit`?