Start using tarfile.data_filter as soon as possible
tarfile.TarFile.extraction_filter is implemented as a means of
mitigating CVE-2007-4559.
Officially, the extraction filters exist only since python 3.12 and the
`data_filter` will become the default in python 3.14.

In build.py:
Many linux distributions backported this patch to earlier versions of
python. Possibly other operating systems have done the same. Thus we
attempt to switch to the `data_filter` regardless of version and
catch the `AttributeError`.

In update_clang_headers.py:
Since this script is not supposed to be run by users, it is fine to
expect maintainers to have a patched/new enough python for updating the
clang headers.