Move DocumentSymbol support into C# completer
We have never advertised the DocumentSymbol capability.
OmniSharp-Roslyn is not respecting that and forcing us to implement it
anyway.
At the same time, we want a flat response, so, with DocumentSymbol, we
need to flatten the hierarchy ourselves.
Since that comes at a performance hit, only C# is handling that and
we're still not claiming DocumentSymbol capability.